check_color_blindness {prismatic}	R Documentation

Visualize color vision deficiency

Description

Visualize color vision deficiency

Usage

check_color_blindness(col)

Arguments

col
a color object or vector of any of the three kinds of R color specifications, i.e., either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a positive integer i meaning palette()[i].

This function will showcase the effect of all 3 kinds of color vision deficiency at the same time side by side.

Value

Nothing

Examples

check_color_blindness(rainbow(10))

check_color_blindness(terrain.colors(10))

[Package prismatic version 1.1.2 Index]