Extract Multiple Components

Extract multiple color components at the same time.

Usage

clr_extract(
  col,
  components = c("red", "green", "blue", "hue_hsl", "saturation", "lightness", "hue_hcl",
    "chroma", "luminance")
)

Arguments

col

a color object or vector of any of the three kinds of R color specifications, i.e., either a color name (as listed by colors()), a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb), or a positive integer i meaning palette()[i].

components

character, components that should be extracted. See details for allowed components.

Value

data.frame of components

Details

The allowed values for `components` are

- red - green - blue - hue_hsl - saturation - lightness - hue_hcl - chroma - luminance

This function is to be preferred if you need to extract multiple components at the same time, since it doesn't need repeat transformations.

See also

Examples

clr_extract(rainbow(10))
#>    red green blue hue_hsl saturation lightness   hue_hcl    chroma luminance
#> 1  255     0    0       0        100        50  12.17399 179.04136  53.24079
#> 2  255   153    0      36        100        50  39.43502 109.90318  72.26072
#> 3  204   255    0      72        100        50 102.90733 109.54895  93.60533
#> 4   51   255    0     108        100        50 126.52222 133.56783  88.07403
#> 5    0   255  102     144        100        50 132.97302 118.64967  88.19634
#> 6    0   255  255     180        100        50 192.17397  72.09871  91.11322
#> 7    0   102  255     216        100        50 259.85568 127.33057  47.90478
#> 8   51     0  255     252        100        50 267.22808 132.04755  33.81896
#> 9  204     0  255     288        100        50 291.97754 133.06693  51.90416
#> 10 255     0  153     324        100        50 346.98632 133.87695  55.65103

clr_extract(rainbow(10), c("hue_hsl", "saturation"))
#>    hue_hsl saturation
#> 1        0        100
#> 2       36        100
#> 3       72        100
#> 4      108        100
#> 5      144        100
#> 6      180        100
#> 7      216        100
#> 8      252        100
#> 9      288        100
#> 10     324        100